// frontend/src/types/user.ts
// src/types/user.ts

export interface User {
    user_id: number;
    name: string;
    phone: string;
    email: string;
    age: number;
    gender: string;
    Netflix: number;
    Amazon_Prime: number;
    DisneyPlus: number;
    ParamountPlus: number;
    Max: number;
    Hulu: number;
    AppleTVPlus: number;
    Peacock: number;
    city: string;
    state: string;
    zip: number;
  }
  